Rework FolderSerializer to use native DRF validation
FullCleanModelSerializer does not work properly with internally-set fields
like "tree", since the value is not available at the time of initial
validation. Currently, it only works because "tree" has "editable=False",
so the Model-layer validator skips the field.

This refactor also has the result of effectively removing the default of
"parent=None" when creating a folder. This is probably a net benefit, as it
forces API clients to be explicit when creating a new root folder.